Refactor exercise and exercise-service in main-frontend
