+++ services/main-frontend/src/components/page-specific/manage/exercise-services/ContentArea.tsx
