+++ services/main-frontend/src/utils/convert.ts
@@ -1,0 +1,3 @@
+export const convertToSlug = (name: string) => {
+  return name.toLowerCase().trim().replaceAll(" ", "-")
+}
